chore: 캐릭터 컴포넌트 분리
